firmware: Store all global variables in registers
This saves over 200 bytes of flash space. Most important savings are
because variables no longer need to be loaded and saved to memory, which
saves a few bytes on every variable access. Additionally, because there
are no longer any globals in memory, the "clear bss" loop is no longer
generated by the compiler, saving a few more bytes.

GCC does have a bug related to global register variables, for which a
workaround is added when using GCC < 4.8 (the bug is fixed in 4.8).