firmware: Add custom "prologues" to two ISRs
This removes any variable (or at least, compiler-dependent) delay from
the INT0 and TIM0_COMPA interrupt handlers with regard to their timing
critical parts.

Note that GCC doesn't actually support adding a custom prologue before
the regular prologue in a interrupt handler, so we use naked interrupt
handlers which then jump to a second function which is also defined as
an interrupt handler. This second function takes care of the real (non
timing-critical) work and also automatically saves any registers it
clobbers.