firmware: Move the mute handling back to the INT0 ISR
When the main loop updates the current action, FLAG_MUTE would not be
respected (since it would have been applied at the end of the previous
bit). In practice, this doesn't really matter, since the main loop will
only set action to ACTION_READY or ACTION_IDLE, neither of which are
affected by FLAG_MUTE. However, it still makes more sense to change
this, to also prevent problems on future changes.