firmware: Also use timer0 overflow interrupt
This allows splitting the timer events into three separate interrupts:
output compare A, output compare B and overflow which respectively
handle sampling the bus, releasing the bus and detecting reset.

An advantage of this approach is the "next action" handling that happens
in the COMPA ISR does not have to wait to the end of the bit time for
a low bit to be released, but can instead always happen either at the
start of the bit, or after sampling the bus halfway. In particular this
should allow the processing (reading EEPROM etc.) between bytes to
happen without having to stall the bus sometimes (which was previously
the case).