firmware: Fix another race condition
When applying a high-speed signal to the bus (say, the programming of
another slave), some assumptions failed to hold. It turned out the INT0
interrupt was triggered while the line was low, which caused the compare
B interrupt to be disabled, leaving the line forever low.

This is expected if there is a falling edge between the start of the
INT0 ISR and the point where it starts pulling the bus low. However,
even clearing the INTF0 flag _after_ pulling the line low (even
with a few cycles delay) did not seem to fix this, so perhaps there is
something else going on.

In any case, this fix just makes sure that whenever the INT0 interrupt
triggers, the line is released, which should prevent the deadlock and
always allows the slave to be reset normally.