firmware: Move the decision about power down mode
Before, the overflow handler would set the sleep mode to power down if
the bus was high. It could do this, since it was certain all timers were
now disabled. Instead, we now just check the TIMSK0 register before
going to sleep mode and use power down mode when no timers are running
(provided the bus is also high, since we can't use an edge-triggered
interrupt to wake from power down).

The most notable effect of this change is that the slave can go into
power down mode at startup, directly, instead of only after the first
falling edge on the bus.