docs/PBBP: Clarify out-of-bounds EEPROM addressing
Previously, the last succesful byte would be nacked, but this was
confusing. The implementation already used the alternative, to send a
dummy byte after the last byte which is then acked. This is slightly
more complicated to implement, but makes a lot more sense.

Additionally, there is no longer a separate error code for end of
EEPROM, just one "Invalid address" error code should be enough.

Closes #5