firmware: Compile for attiny13a instead of attiny13
The generated code is effectively the same and compatible, but attiny13a
has some extra registers we should be using.

Note that there are some gcc/libc bugs with the attiny13a. One
workaround is included in the code, another must be implemented on your
local machine by creating a symlink:

	ln -s /usr/lib/avr/lib/avr25 /usr/lib/avr/lib/avr25/tiny-stack

This wasn't a problem for attiny13, because gcc falls back to
the avr2 architecture without tiny-stack (/usr/lib/avr/lib/), where the
attiny13 also lives (for compatibility with gcc versions that don't know
about avr25, I think, which isn't the case here, it just accidentally
masks the missing tiny-stack problem).