firmware: Enable the brown-out detector
This sets the brownout level at 2.7V (typical). Also, it disables the
brownout detector during power-down, to save power. This implies a bit
of extra startup time, so at the same time power-down mode is only
enabled in idle mode, so we onlly need to detect a reset signal after
waking up and the extra delay doesn't matter much.

Closes #3