firmware: Reduce power-up delay to 14CK
The previous fuse settings were set to delay 14CK + 64ms after power-up
and reset, to allow the power supply to become stable. Now that the
brown-out detector is enabled, it is safe to reduce this delay to 14CK
(since the BOD will further delay the startup when the voltage isn't
high enough yet).