firmware: Prevent triggering INT0 when the slave pulls the line low
Normally, the line is pulled low before the slave releases the line.
However, the protocol does not require this, so when the master is using
minimal timings, it could happen that the line is high before the slave
pulls it low. In this case, we should prevent triggering INT0 a second
time.