firmware-test: Make the sketch work with the IDE
Before, it was only tested using Arduino-Makefile, which doesn't do all
the pre-processing the IDE does. This preprocessing turned out to break
the code, so all code was moved into a .cpp file to prevent it from
being preprocessed like a .ino file. There is still a dummy .ino file,
to make the IDE recognize the directory as a sketch.