docs/EEPROM: Store the used EEPROM size explicitly
Previously, the used part of the EEPROM was defined to end after the
Checksum descriptor. However, this would mean that to determine if the
checksum was correct, the type of all descriptors in the EEPROM should
be read and their length deduced, even before the checksum can be
confirmed.

By storing the length explicitely, the location of the checksum can be
determined by just looking at two bytes (layout version and used size)
and checked easily. Because now the checksum position is known, it no
longer needs to be a checksum descriptor, so the extra byte in the
header is compensated by the removal of the checksum descriptor type
byte.